refactor: move trash os_limited functions to trash_helpers module
Prereq. for https://github.com/pop-os/cosmic-files/pull/1741 : perf: fix slow startup by moving blocking trash checks to async background tasks
These changes were made with the assistance of Gemini 3.1 Pro Preview and devstral-2 (LLMs)
